move make_run_file
