+++ tests/image_stitching_test.py
@@ -5,45 +5,8 @@ import tempfile
 import pytest
 from alpineer import io_utils, load_utils, test_utils
 
-from toffy import image_stitching, json_utils
-
-
-def make_run_file(tmp_dir, prefixes=[]):
-    """Create a run subir and run json in the provided dir and return the path to this new dir."""
-
-    if len(prefixes) == 1:
-        prefix1, prefix2 = prefixes * 2
-    else:
-        prefix1 = prefixes[0]
-        prefix2 = prefixes[1]
-
-    fov_data = {
-        f"{prefix1}R1C3": 32,
-        f"{prefix1}R2C1": 16,
-        f"{prefix1}MoQC": 8,
-        f"{prefix2}R2C2": 16,
-        "nontiled": 16,
-    }
-    run_data = []
-
-    for i, fov in enumerate(fov_data.keys()):
-        run_data.append(
-            {
-                "runOrder": i + 1,
-                "scanCount": 1,
-                "frameSizePixels": {"width": fov_data[fov], "height": fov_data[fov]},
-                "name": fov,
-            }
-        )
-
-    run_json_spoof = {"fovs": run_data}
-
-    test_dir = os.path.join(tmp_dir, "data", "test_run")
-    os.makedirs(test_dir)
-    json_path = os.path.join(test_dir, "test_run.json")
-    json_utils.write_json_file(json_path, run_json_spoof)
-
-    return test_dir
+from tests.utils.test_utils import make_run_file
+from toffy import image_stitching
 
 
 def test_get_max_img_size():
